+++ python/GafferUI/CompoundDataPlugValueWidget.py
@@ -138,7 +138,7 @@ class CompoundDataPlugValueWidget( GafferUI.PlugValueWidget ) :
 		result.append( "/Add/Box3i", { "command" : functools.partial( Gaffer.WeakMethod( self.__addItem ), "", IECore.Box3iData( imath.Box3i( imath.V3i( 0 ), imath.V3i( 1 ) ) ) ) } )
 		result.append( "/Add/Box3f", { "command" : functools.partial( Gaffer.WeakMethod( self.__addItem ), "", IECore.Box3fData( imath.Box3f( imath.V3f( 0 ), imath.V3f( 1 ) ) ) ) } )
 
-		result.append( "/Add/BoxDivider", { "divider" : True } )
+		result.append( "/Add/ArrayDivider", { "divider" : True } )
 
 		for label, plugType in [
 			( "Float", Gaffer.FloatVectorDataPlug ),